chore: move loadingContainer to atoms
